consolidated dat readers into one w/ a cli flag
